+++ flaker-android-okhttp/src/androidTest/java/io/rotlabs/flakerandroidretrofit/ExampleInstrumentedTest.kt
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("io.rotlabs.flakerandroidretrofit.test", appContext.packageName)
+        assertEquals("io.rotlabs.flakerandroidokhttp.test", appContext.packageName)
     }
 }
