+++ flaker-android-okhttp/src/main/AndroidManifest.xml
@@ -1,14 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="io.rotlabs.flakerandroidretrofit">
+    package="io.rotlabs.flakerandroidokhttp">
 
     <application
         android:allowBackup="true"
         android:supportsRtl="true"
         tools:targetApi="33">
         <activity
-            android:name="io.rotlabs.flakerandroidretrofit.ui.FlakerActivity"
+            android:name=".ui.FlakerActivity"
             android:exported="true"
             android:icon="@mipmap/ic_flaker"
             android:label="@string/companion_app_name"
