+++ flaker-android-okhttp/src/main/java/io/rotlabs/flakerandroidokhttp/FlakerIntentAction.kt
@@ -1,0 +1,3 @@
+package io.rotlabs.flakerandroidokhttp
+
+const val FLAKER_ACTION_OPEN = "io.rotlabs.flakerandroidokhttp.OPEN"
