--- flaker-android-okhttp/src/main/java/io/rotlabs/flakerandroidretrofit/FlakerIntentAction.kt
@@ -1,3 +1,0 @@
-package io.rotlabs.flakerandroidretrofit
-
-const val FLAKER_ACTION_OPEN = "io.rotlabs.flakerandroidretrofit.OPEN"
