+++ flaker-okhttp-core/src/androidTest/java/io/rotlabs/flakerretrofit/ExampleInstrumentedTest.kt
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("io.rotlabs.flakerretrofit.test", appContext.packageName)
+        assertEquals("io.rotlabs.flakerokhttp.test", appContext.packageName)
     }
 }
