+++ flaker-okhttp-core/src/test/java/io/rotlabs/flakerokhttpcore/TestMonitorContainer.kt
@@ -1,4 +1,4 @@
-package io.rotlabs.flakerretrofit
+package io.rotlabs.flakerokhttpcore
 
 import android.content.Context
 import io.rotlabs.flakerandroidmonitor.FlakerMonitor
